% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_load_data.R
\name{get_mass_spec_basic_tof_files}
\alias{get_mass_spec_basic_tof_files}
\title{get_mass_spec_basic_tof_files}
\usage{
get_mass_spec_basic_tof_files()
}
\value{
A character vector of absolute file paths.
}
\description{
Gets full paths of basic TOF files (centroid mzML, centroid
mzXML, and profile mzML formats).
}
