% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_load_data.R
\name{get_mass_spec_basic_tof_suspects_csv}
\alias{get_mass_spec_basic_tof_suspects_csv}
\title{get_mass_spec_basic_tof_suspects_csv}
\usage{
get_mass_spec_basic_tof_suspects_csv()
}
\value{
A character string with the absolute file path.
}
\description{
Gets the path to the basic TOF suspects CSV file containing
target compounds for screening.
}
