% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_load_data.R
\name{get_mass_spec_files}
\alias{get_mass_spec_files}
\title{get_mass_spec_files}
\usage{
get_mass_spec_files()
}
\value{
A character vector of absolute file paths.
}
\description{
Gets full paths of all mass spectrometry files (mzML, mzXML)
in the StreamFindData package.
}
