% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_load_data.R
\name{get_mass_spec_mrm_estrogens_csv}
\alias{get_mass_spec_mrm_estrogens_csv}
\title{get_mass_spec_mrm_estrogens_csv}
\usage{
get_mass_spec_mrm_estrogens_csv()
}
\value{
A character string with the absolute file path.
}
\description{
Gets the path to the MRM estrogens CSV file with target and
product m/z values for estrogenic compounds.
}
