% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_load_data.R
\name{get_mass_spec_mrm_estrogens_files}
\alias{get_mass_spec_mrm_estrogens_files}
\title{get_mass_spec_mrm_estrogens_files}
\usage{
get_mass_spec_mrm_estrogens_files()
}
\value{
A character vector of absolute file paths.
}
\description{
Gets full paths of MRM estrogens files (negative polarity).
}
