% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_load_data.R
\name{get_mass_spec_mrm_nitrosamines_files}
\alias{get_mass_spec_mrm_nitrosamines_files}
\title{get_mass_spec_mrm_nitrosamines_files}
\usage{
get_mass_spec_mrm_nitrosamines_files()
}
\value{
A character vector of absolute file paths.
}
\description{
Gets full paths of MRM nitrosamines files (positive polarity).
}
