% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_load_data.R
\name{get_mass_spec_wastewater_files}
\alias{get_mass_spec_wastewater_files}
\title{get_mass_spec_wastewater_files}
\usage{
get_mass_spec_wastewater_files()
}
\value{
A character vector of absolute file paths.
}
\description{
Gets full paths of wastewater project TOF-MS files (blanks,
influent, and ozonated effluent in positive and negative polarity).
}
