% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_load_data.R
\name{get_mass_spec_wastewater_internal_standards_csv}
\alias{get_mass_spec_wastewater_internal_standards_csv}
\title{get_mass_spec_wastewater_internal_standards_csv}
\usage{
get_mass_spec_wastewater_internal_standards_csv()
}
\value{
A character string with the absolute file path.
}
\description{
Gets the path to the wastewater internal standards CSV file.
}
