% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_load_data.R
\name{get_raman_files}
\alias{get_raman_files}
\title{get_raman_files}
\usage{
get_raman_files()
}
\value{
A character vector of absolute file paths.
}
\description{
Gets full paths of Raman spectroscopy .asc files (Bevacizumab
drug product and blank measurements).
}
