% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamfind.data-package.R
\docType{package}
\name{streamfind.data-package}
\alias{streamfind.data}
\alias{streamfind.data-package}
\title{streamfind.data: Package to hold supporting data for package streamFind.}
\description{
Package to hold supporting data for package streamFind..
}
\keyword{internal}
